Fixed image processing to omit blank values
I found that my image processing results were being heavily skewed by the background of the image. When counting the pixels of an image, the background was being counted as black  which would lead to very high counts in the black pixel category. I have fixed this to omit over 60,000 black pixels for each image(~half of the pixels in the image) and will be testing this image dataset to see if results improve.